Save coords separately
