+++ simulation_test.py
@@ -1,4 +1,4 @@
-from TkPixels.Design import Design
+from TkPixels.StripDesign import Design
 
 strip = Design(60, 0.8)
 strip.fill((255,0,0))
